Reapply "Added a /redis and /ioredis export path (#9)" (#10)
This reverts commit 57a5815c6e42ad5fdf3bcf09346d377dbb291c4b.
