+++ docs/README.md
@@ -4,6 +4,10 @@
 
 # Resumable Stream v2.1.0
 
+## Namespaces
+
+- [\_Private](Resumable-Stream/namespaces/Private/README.md)
+
 ## Interfaces
 
 - [CreateResumableStreamContextOptions](interfaces/CreateResumableStreamContextOptions.md)
@@ -11,7 +15,10 @@
 - [ResumableStreamContext](interfaces/ResumableStreamContext.md)
 - [Subscriber](interfaces/Subscriber.md)
 
+## Variables
+
+- [createResumableStreamContext](variables/createResumableStreamContext.md)
+
 ## Functions
 
-- [createResumableStreamContext](functions/createResumableStreamContext.md)
 - [resumeStream](functions/resumeStream.md)
