+++ docs/Resumable-Stream/namespaces/Private/README.md
@@ -1,0 +1,11 @@
+[**Resumable Stream v2.1.0**](../../../README.md)
+
+***
+
+[Resumable Stream](../../../README.md) / \_Private
+
+# \_Private
+
+## Type Aliases
+
+- [RedisDefaults](type-aliases/RedisDefaults.md)
