+++ docs/Resumable-Stream/namespaces/Private/type-aliases/RedisDefaults.md
@@ -1,0 +1,33 @@
+[**Resumable Stream v2.1.0**](../../../../README.md)
+
+***
+
+[Resumable Stream](../../../../README.md) / [\_Private](../README.md) / RedisDefaults
+
+# Type Alias: RedisDefaults
+
+> **RedisDefaults** = `object`
+
+## Properties
+
+### publisher()
+
+> **publisher**: () => [`Publisher`](../../../../interfaces/Publisher.md)
+
+A pubsub publisher. Designed to be compatible with clients from the `redis` package.
+
+#### Returns
+
+[`Publisher`](../../../../interfaces/Publisher.md)
+
+***
+
+### subscriber()
+
+> **subscriber**: () => [`Subscriber`](../../../../interfaces/Subscriber.md)
+
+A pubsub subscriber. Designed to be compatible with clients from the `redis` package.
+
+#### Returns
+
+[`Subscriber`](../../../../interfaces/Subscriber.md)
