+++ src/get-redis-url.ts
@@ -1,0 +1,7 @@
+export function getRedisUrl() {
+  const redisUrl = process.env.REDIS_URL || process.env.KV_URL;
+  if (!redisUrl) {
+    throw new Error("REDIS_URL environment variable is not set");
+  }
+  return redisUrl;
+}
